requirements.txt
torch
transformers
Pillow
scikit-learn
